pyproject.toml
[project]
name = "New_AI_Kanilla"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=1.52.0",
    "ipykernel>=7.1.0",
    "lancedb>=0.26.0",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pydantic-ai>=1.39.0",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.2.0",
    "uvicorn>=0.40.0",
]


